# docs/api-document.yaml
openapi: "3.0.2"
info:
  title: "kininaru_clip_API"
  version: "2.0.0"
servers:
  - url: "http://localhost:8080"
  - url: "http://localhost:3000"
paths:
...
  /events:
    post:
      tags:
        - "events"
      summary: "Create a new event"
      operationId: "createEvent"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/createEventInput"
      responses:
        "200":
          description: "Event created"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/createEventResponse"
...
components:
  schemas:
    createEventInput:
      type: object
      properties:
        title:
          type: string
          description: "イベントのタイトル"
        url:
          type: "string"
        users:
          type: "array"
          description: "作成するユーザーの入力情報の配列"
          items:
            type: object
            properties:
              name:
                type: string
                description: "ユーザーの名前"
            required:
              - name
      required:
        - "title"
        - "url"
        - "users"
    createEventResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: "イベントの一意な識別子"
        title:
          type: string
          description: "イベントのタイトル"
        url:
          type: "string"
          description: "イベントのURL"
      required:
        - "id"
        - "title"
        - "url"
...